feat(look&feel,apollo): css optimization & cell size modification
