--- apps/look-and-feel-stories/src/components/Table/TableAdvanced.stories.scss
@@ -1,17 +1,0 @@
-.container {
-    width: 100%;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    
-}
-
-.dropdown {
-    max-width: 420px;
-}
-
-.pagination {
-    justify-items: center;
-    width: 100%;
-}
